+++ app/modules/event/provides.php
@@ -1,7 +1,0 @@
-<?php
-
-return array(
-    'widgets' => array(
-        'messages' => array('view' => 'messages_widget'),
-    ),
-);