+++ public/index.php
@@ -24,17 +24,21 @@ function fatal($msg)
     exit(1);
 }
 
+
 //===============================================
 // Autoloading
 //===============================================
+require_once APP_ROOT.'app/helpers/env_helper.php';
 if ((include APP_ROOT . "vendor/autoload.php") === false)
 {
     fatal("vendor/autoload.php is missing!<br>
 Please run `composer install` in the munkireport directory</p>");
 }
 
-// Load config
+// Include config helper
 require APP_ROOT.'app/helpers/config_helper.php';
+
+// Load config
 initDotEnv();
 load_conf();